feat: consolidate server-side functions and restructure architecture for clearer separation of client and server code; update imports and documentation accordingly
