+++ src/components/movies/MovieCard.tsx
@@ -1,13 +1,11 @@
-"use client";
-
 import Image from "next/image";
-import { useRouter } from "next/navigation";
-import { useState, useEffect } from "react";
 import { TMDBMovie } from "@/lib/tmdb";
 import { getPosterUrl } from "@/lib/movieUtils";
+import { MovieCardClient } from "./MovieCardClient";
 
-// MOVIE CARD: Individual movie display component
-// This component displays a single movie with poster, title, and rating
+// SERVER COMPONENT: Static movie card content
+// This component handles all server-side rendering (SEO, static content)
+// No "use client" directive - runs on server
 interface MovieCardProps {
   movie: TMDBMovie;
   isLoading?: boolean;
@@ -29,24 +27,6 @@ export const MovieCard = ({
   isRatingLoading = false,
   isWantToWatchLoading = false,
 }: MovieCardProps) => {
-  const router = useRouter();
-
-  // Add client-side only state to prevent hydration mismatches
-  const [isClient, setIsClient] = useState(false);
-
-  // Set client flag after hydration
-  useEffect(() => {
-    setIsClient(true);
-  }, []);
-
-  const handleMovieClick = () => {
-    // Set the current page path in sessionStorage before navigating
-    if (typeof window !== "undefined") {
-      sessionStorage.setItem("previousPage", window.location.pathname);
-    }
-    router.push(`/movies/${movie.id}`);
-  };
-
   if (isLoading) {
     return (
       <div className="bg-white dark:bg-slate-800 rounded-lg shadow-sm border border-slate-200 dark:border-slate-700 overflow-hidden animate-pulse">
@@ -61,121 +41,26 @@ export const MovieCard = ({
 
   return (
     <div className="bg-white dark:bg-slate-800 rounded-lg shadow-sm border border-slate-200 dark:border-slate-700 overflow-hidden hover:shadow-md transition-shadow group">
-      <div
-        onClick={handleMovieClick}
-        className="w-full text-left cursor-pointer"
-      >
-        <div className="relative aspect-[2/3] overflow-hidden">
-          <Image
-            src={getPosterUrl(movie.poster_path)}
-            alt={movie.title}
-            fill
-            className="object-cover transition-transform group-hover:scale-105"
-            sizes="(max-width: 768px) 50vw, (max-width: 1200px) 33vw, 25vw"
-          />
-          {isClient && onOpenRatingModal && (
-            <button
-              onClick={(e) => {
-                e.stopPropagation();
-                onOpenRatingModal();
-              }}
-              disabled={isRatingLoading}
-              className={`absolute top-2 left-2 p-1.5 rounded-full transition-colors ${
-                isRatingLoading
-                  ? "bg-gray-400 cursor-not-allowed"
-                  : userRating
-                  ? "bg-yellow-500 hover:bg-yellow-600 text-white"
-                  : "bg-white/80 hover:bg-white text-gray-700 hover:text-yellow-600"
-              }`}
-            >
-              {isRatingLoading ? (
-                <svg
-                  className="animate-spin h-4 w-4"
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                >
-                  <circle
-                    className="opacity-25"
-                    cx="12"
-                    cy="12"
-                    r="10"
-                    stroke="currentColor"
-                    strokeWidth="4"
-                  ></circle>
-                  <path
-                    className="opacity-75"
-                    fill="currentColor"
-                    d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-                  ></path>
-                </svg>
-              ) : userRating ? (
-                <span className="font-bold flex items-center justify-center w-4 h-4">
-                  {userRating}
-                </span>
-              ) : (
-                <svg
-                  className="h-4 w-4 fill-none"
-                  stroke="currentColor"
-                  strokeWidth="2"
-                  viewBox="0 0 24 24"
-                >
-                  <path d="M12 2l3.09 6.26L22 9.27l-5 4.87 1.18 6.88L12 17.77l-6.18 3.25L7 14.14 2 9.27l6.91-1.01L12 2z" />
-                </svg>
-              )}
-            </button>
-          )}
-          {isClient && onToggleWantToWatch && (
-            <button
-              onClick={(e) => {
-                e.stopPropagation();
-                onToggleWantToWatch(movie, isInWantToWatch);
-              }}
-              disabled={isWantToWatchLoading}
-              className={`absolute top-2 right-2 p-1.5 rounded-full transition-colors ${
-                isWantToWatchLoading
-                  ? "bg-gray-400 cursor-not-allowed"
-                  : isInWantToWatch
-                  ? "bg-red-500 hover:bg-red-600 text-white"
-                  : "bg-white/80 hover:bg-white text-gray-700 hover:text-red-500"
-              }`}
-            >
-              {isWantToWatchLoading ? (
-                <svg
-                  className="animate-spin h-4 w-4"
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                >
-                  <circle
-                    className="opacity-25"
-                    cx="12"
-                    cy="12"
-                    r="10"
-                    stroke="currentColor"
-                    strokeWidth="4"
-                  ></circle>
-                  <path
-                    className="opacity-75"
-                    fill="currentColor"
-                    d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-                  ></path>
-                </svg>
-              ) : (
-                <svg
-                  className={`h-4 w-4 ${
-                    isInWantToWatch ? "fill-current" : "fill-none"
-                  }`}
-                  stroke="currentColor"
-                  strokeWidth="2"
-                  viewBox="0 0 24 24"
-                >
-                  <path d="M20.84 4.61a5.5 5.5 0 0 0-7.78 0L12 5.67l-1.06-1.06a5.5 5.5 0 0 0-7.78 7.78l1.06 1.06L12 21.23l7.78-7.78 1.06-1.06a5.5 5.5 0 0 0 0-7.78z" />
-                </svg>
-              )}
-            </button>
-          )}
-        </div>
+      {/* Static content - renders on server */}
+      <div className="relative aspect-[2/3] overflow-hidden">
+        <Image
+          src={getPosterUrl(movie.poster_path)}
+          alt={movie.title}
+          fill
+          className="object-cover transition-transform group-hover:scale-105"
+          sizes="(max-width: 768px) 50vw, (max-width: 1200px) 33vw, 25vw"
+        />
+
+        {/* Client component handles all interactive elements */}
+        <MovieCardClient
+          movie={movie}
+          userRating={userRating}
+          isInWantToWatch={isInWantToWatch}
+          onOpenRatingModal={onOpenRatingModal}
+          onToggleWantToWatch={onToggleWantToWatch}
+          isRatingLoading={isRatingLoading}
+          isWantToWatchLoading={isWantToWatchLoading}
+        />
       </div>
 
       <div className="p-3">
