--- src/hooks/server/index.ts
@@ -1,7 +1,0 @@
-// Server-side data functions
-export {
-  getInitialMovies,
-  getMovieData,
-  getInitialGenres,
-  getMoviesByGenreData,
-} from "./useServerData";
